requirements.txt
Flask
Pillow
matplotlib
gunicorn
pytest
